requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pyyaml==6.0.1
schedule==1.2.0
jinja2==3.1.2
pandas==2.1.0
python-dotenv==1.0.0
fake-useragent==1.4.0
selenium==4.15.0
webdriver-manager==4.0.1
